Fixed small issue with the level transition

+++ STMSerial/Core/Inc/Entity.h
@@ -18,8 +18,8 @@
 //	}
 //};
 struct pointVector {
-	uint8_t X;
-	uint8_t Y;
+	int X;
+	int Y;
 
 };
 class Entity
